Move project CRUD operations to a separate file
